Move sr-arithmetic to its own crate
