+++ Cargo.toml
@@ -53,6 +53,7 @@ members = [
 	"core/service/test",
 	"core/session",
 	"core/sr-api-macros",
+	"core/sr-arithmetic",
 	"core/sr-io",
 	"core/sr-primitives",
 	"core/sr-staking-primitives",
