+++ core/sr-arithmetic/Cargo.toml
@@ -1,21 +1,17 @@
 [package]
-name = "sr-primitives"
+name = "sr-arithmetic"
 version = "2.0.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2018"
 
 [dependencies]
+num-traits = { version = "0.2.8", default-features = false }
 serde = { version = "1.0.101", optional = true, features = ["derive"] }
 codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
-primitives = { package = "substrate-primitives",  path = "../primitives", default-features = false }
-app-crypto = { package = "substrate-application-crypto",  path = "../application-crypto", default-features = false }
-arithmetic = { package = "sr-arithmetic",  path = "../sr-arithmetic", default-features = false }
 rstd = { package = "sr-std", path = "../sr-std", default-features = false }
-runtime_io = { package = "sr-io", path = "../sr-io", default-features = false }
-log = { version = "0.4.8", optional = true }
 paste = "0.1.6"
 rand = { version = "0.7.2", optional = true }
-impl-trait-for-tuples = "0.1.2"
+integer-sqrt = "0.1.2"
 
 [dev-dependencies]
 serde_json = "1.0.41"
@@ -27,13 +23,9 @@ substrate-offchain = { path = "../offchain" }
 bench = []
 default = ["std"]
 std = [
-	"serde",
-	"log",
+  "serde",
+	"num-traits/std",
 	"rstd/std",
-	"runtime_io/std",
 	"codec/std",
-	"primitives/std",
-	"app-crypto/std",
-	"arithmetic/std",
 	"rand",
 ]
