+++ core/sr-primitives/src/traits.rs
@@ -17,28 +17,20 @@
 //! Primitives for the runtime modules.
 
 use rstd::prelude::*;
-use rstd::{self, result, marker::PhantomData, convert::{TryFrom, TryInto}};
+use rstd::{self, result, marker::PhantomData};
 use runtime_io;
 #[cfg(feature = "std")]
 use std::fmt::{Debug, Display};
 #[cfg(feature = "std")]
 use serde::{Serialize, Deserialize, de::DeserializeOwned};
 use primitives::{self, Hasher, Blake2Hasher, TypeId};
-use crate::codec::{Codec, Encode, Decode, HasCompact};
+use crate::codec::{Codec, Encode, Decode};
 use crate::transaction_validity::{
 	ValidTransaction, TransactionValidity, TransactionValidityError, UnknownTransaction,
 };
+pub use arithmetic::traits::*;
 use crate::generic::{Digest, DigestItem};
 use crate::weights::DispatchInfo;
-pub use integer_sqrt::IntegerSquareRoot;
-pub use num_traits::{
-	Zero, One, Bounded, CheckedAdd, CheckedSub, CheckedMul, CheckedDiv,
-	CheckedShl, CheckedShr
-};
-use rstd::ops::{
-	Add, Sub, Mul, Div, Rem, AddAssign, SubAssign, MulAssign, DivAssign,
-	RemAssign, Shl, Shr
-};
 use app_crypto::AppKey;
 use impl_trait_for_tuples::impl_for_tuples;
 
@@ -192,189 +184,6 @@ impl<T> Convert<T, T> for Identity {
 	fn convert(a: T) -> T { a }
 }
 
-/// A structure that performs standard conversion using the standard Rust conversion traits.
-pub struct ConvertInto;
-impl<A, B: From<A>> Convert<A, B> for ConvertInto {
-	fn convert(a: A) -> B { a.into() }
-}
-
-/// A meta trait for arithmetic.
-///
-/// Arithmetic types do all the usual stuff you'd expect numbers to do. They are guaranteed to
-/// be able to represent at least `u32` values without loss, hence the trait implies `From<u32>`
-/// and smaller ints. All other conversions are fallible.
-pub trait SimpleArithmetic:
-	Zero + One + IntegerSquareRoot +
-	From<u8> + From<u16> + From<u32> + TryInto<u8> + TryInto<u16> + TryInto<u32> +
-	TryFrom<u64> + TryInto<u64> + TryFrom<u128> + TryInto<u128> + TryFrom<usize> + TryInto<usize> +
-	UniqueSaturatedInto<u8> + UniqueSaturatedInto<u16> + UniqueSaturatedInto<u32> +
-	UniqueSaturatedFrom<u64> + UniqueSaturatedInto<u64> + UniqueSaturatedFrom<u128> + UniqueSaturatedInto<u128> +
-	Add<Self, Output = Self> + AddAssign<Self> +
-	Sub<Self, Output = Self> + SubAssign<Self> +
-	Mul<Self, Output = Self> + MulAssign<Self> +
-	Div<Self, Output = Self> + DivAssign<Self> +
-	Rem<Self, Output = Self> + RemAssign<Self> +
-	Shl<u32, Output = Self> + Shr<u32, Output = Self> +
-	CheckedShl + CheckedShr + CheckedAdd + CheckedSub + CheckedMul + CheckedDiv +
-	Saturating + PartialOrd<Self> + Ord + Bounded +
-	HasCompact + Sized
-{}
-impl<T:
-	Zero + One + IntegerSquareRoot +
-	From<u8> + From<u16> + From<u32> + TryInto<u8> + TryInto<u16> + TryInto<u32> +
-	TryFrom<u64> + TryInto<u64> + TryFrom<u128> + TryInto<u128> + TryFrom<usize> + TryInto<usize> +
-	UniqueSaturatedInto<u8> + UniqueSaturatedInto<u16> + UniqueSaturatedInto<u32> +
-	UniqueSaturatedFrom<u64> + UniqueSaturatedInto<u64> + UniqueSaturatedFrom<u128> +
-	UniqueSaturatedInto<u128> + UniqueSaturatedFrom<usize> + UniqueSaturatedInto<usize> +
-	Add<Self, Output = Self> + AddAssign<Self> +
-	Sub<Self, Output = Self> + SubAssign<Self> +
-	Mul<Self, Output = Self> + MulAssign<Self> +
-	Div<Self, Output = Self> + DivAssign<Self> +
-	Rem<Self, Output = Self> + RemAssign<Self> +
-	Shl<u32, Output = Self> + Shr<u32, Output = Self> +
-	CheckedShl + CheckedShr + CheckedAdd + CheckedSub + CheckedMul + CheckedDiv +
-	Saturating + PartialOrd<Self> + Ord + Bounded +
-	HasCompact + Sized
-> SimpleArithmetic for T {}
-
-/// Just like `From` except that if the source value is too big to fit into the destination type
-/// then it'll saturate the destination.
-pub trait UniqueSaturatedFrom<T: Sized>: Sized {
-	/// Convert from a value of `T` into an equivalent instance of `Self`.
-	fn unique_saturated_from(t: T) -> Self;
-}
-
-/// Just like `Into` except that if the source value is too big to fit into the destination type
-/// then it'll saturate the destination.
-pub trait UniqueSaturatedInto<T: Sized>: Sized {
-	/// Consume self to return an equivalent value of `T`.
-	fn unique_saturated_into(self) -> T;
-}
-
-impl<T: Sized, S: TryFrom<T> + Bounded + Sized> UniqueSaturatedFrom<T> for S {
-	fn unique_saturated_from(t: T) -> Self {
-		S::try_from(t).unwrap_or_else(|_| Bounded::max_value())
-	}
-}
-
-impl<T: Bounded + Sized, S: TryInto<T> + Sized> UniqueSaturatedInto<T> for S {
-	fn unique_saturated_into(self) -> T {
-		self.try_into().unwrap_or_else(|_| Bounded::max_value())
-	}
-}
-
-/// Simple trait to use checked mul and max value to give a saturated mul operation over
-/// supported types.
-pub trait Saturating {
-	/// Saturated addition - if the product can't fit in the type then just use max-value.
-	fn saturating_add(self, o: Self) -> Self;
-
-	/// Saturated subtraction - if the product can't fit in the type then just use max-value.
-	fn saturating_sub(self, o: Self) -> Self;
-
-	/// Saturated multiply - if the product can't fit in the type then just use max-value.
-	fn saturating_mul(self, o: Self) -> Self;
-}
-
-impl<T: CheckedMul + Bounded + num_traits::Saturating> Saturating for T {
-	fn saturating_add(self, o: Self) -> Self {
-		<Self as num_traits::Saturating>::saturating_add(self, o)
-	}
-	fn saturating_sub(self, o: Self) -> Self {
-		<Self as num_traits::Saturating>::saturating_sub(self, o)
-	}
-	fn saturating_mul(self, o: Self) -> Self {
-		self.checked_mul(&o).unwrap_or_else(Bounded::max_value)
-	}
-}
-
-/// Convenience type to work around the highly unergonomic syntax needed
-/// to invoke the functions of overloaded generic traits, in this case
-/// `SaturatedFrom` and `SaturatedInto`.
-pub trait SaturatedConversion {
-	/// Convert from a value of `T` into an equivalent instance of `Self`.
-	///
-	/// This just uses `UniqueSaturatedFrom` internally but with this
-	/// variant you can provide the destination type using turbofish syntax
-	/// in case Rust happens not to assume the correct type.
-	fn saturated_from<T>(t: T) -> Self where Self: UniqueSaturatedFrom<T> {
-		<Self as UniqueSaturatedFrom<T>>::unique_saturated_from(t)
-	}
-
-	/// Consume self to return an equivalent value of `T`.
-	///
-	/// This just uses `UniqueSaturatedInto` internally but with this
-	/// variant you can provide the destination type using turbofish syntax
-	/// in case Rust happens not to assume the correct type.
-	fn saturated_into<T>(self) -> T where Self: UniqueSaturatedInto<T> {
-		<Self as UniqueSaturatedInto<T>>::unique_saturated_into(self)
-	}
-}
-impl<T: Sized> SaturatedConversion for T {}
-
-/// Convenience type to work around the highly unergonomic syntax needed
-/// to invoke the functions of overloaded generic traits, in this case
-/// `TryFrom` and `TryInto`.
-pub trait CheckedConversion {
-	/// Convert from a value of `T` into an equivalent instance of `Option<Self>`.
-	///
-	/// This just uses `TryFrom` internally but with this
-	/// variant you can provide the destination type using turbofish syntax
-	/// in case Rust happens not to assume the correct type.
-	fn checked_from<T>(t: T) -> Option<Self> where Self: TryFrom<T> {
-		<Self as TryFrom<T>>::try_from(t).ok()
-	}
-	/// Consume self to return `Some` equivalent value of `Option<T>`.
-	///
-	/// This just uses `TryInto` internally but with this
-	/// variant you can provide the destination type using turbofish syntax
-	/// in case Rust happens not to assume the correct type.
-	fn checked_into<T>(self) -> Option<T> where Self: TryInto<T> {
-		<Self as TryInto<T>>::try_into(self).ok()
-	}
-}
-impl<T: Sized> CheckedConversion for T {}
-
-/// Multiply and divide by a number that isn't necessarily the same type. Basically just the same
-/// as `Mul` and `Div` except it can be used for all basic numeric types.
-pub trait Scale<Other> {
-	/// The output type of the product of `self` and `Other`.
-	type Output;
-
-	/// @return the product of `self` and `other`.
-	fn mul(self, other: Other) -> Self::Output;
-
-	/// @return the integer division of `self` and `other`.
-	fn div(self, other: Other) -> Self::Output;
-
-	/// @return the modulo remainder of `self` and `other`.
-	fn rem(self, other: Other) -> Self::Output;
-}
-macro_rules! impl_scale {
-	($self:ty, $other:ty) => {
-		impl Scale<$other> for $self {
-			type Output = Self;
-			fn mul(self, other: $other) -> Self::Output { self * (other as Self) }
-			fn div(self, other: $other) -> Self::Output { self / (other as Self) }
-			fn rem(self, other: $other) -> Self::Output { self % (other as Self) }
-		}
-	}
-}
-impl_scale!(u128, u128);
-impl_scale!(u128, u64);
-impl_scale!(u128, u32);
-impl_scale!(u128, u16);
-impl_scale!(u128, u8);
-impl_scale!(u64, u64);
-impl_scale!(u64, u32);
-impl_scale!(u64, u16);
-impl_scale!(u64, u8);
-impl_scale!(u32, u32);
-impl_scale!(u32, u16);
-impl_scale!(u32, u8);
-impl_scale!(u16, u16);
-impl_scale!(u16, u8);
-impl_scale!(u8, u8);
 
 /// Trait for things that can be clear (have no bits set). For numeric types, essentially the same
 /// as `Zero`.
